requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
pymongo
scikit-learn
python-dotenv

